+++ NEWS.md
@@ -8,6 +8,9 @@
 * geom_phylopic now properly handles a single unlisted image object passed to the "img" parameter (#75)
 * added filter (license) argument to get_uuid, pick_phylo, add_phylopic_base, add_phylopic, and geom_phylopic (#72)
 * added img argument to get_uuid and get_attribution
+* split out the functionality of the color argument/aesthetic to color (silhouette outline) and fill (silhouette) arguments/aesthetics in add_phylopic, geom_phylopic, and add_phylopic_base (#58)
+  * when only the color argument/aesthetic is specified, it is copied to the fill argument/aesthetic (maintaining mostly backwards compatibility with old code)
+* added plot and print methods for silhouette objects (#73)
 
 # rphylopic 1.1.1